stock-bot/Flaskrequirements.txt
Flask
line-bot-sdk<3.0.0
google-generativeai
yfinance
gunicorn